autoserved-api-main/autoserved-api-main/requirements.txt
Flask==1.1.2
pandas==1.2.4
numpy==1.20.2
joblib==1.0.1
category_encoders==2.2.2
scikit-multilearn==0.2.0
liac-arff==2.5.0
keras==2.4.3
# ideally tensorflow==2.4.1 or 2.5 it seems to be fine to just run normal tensorflow
python-dotenv==0.17.0
xlrd==1.2.0
requests==2.25.1
waitress==2.0.0
